# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/cv/generate": {
      "post": {
        "operationId": "CvController_generate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/GenerateCvDto" } }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CV" } } }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Generate a CV for a specific job",
        "tags": ["cv"]
      }
    },
    "/api/cv": {
      "get": {
        "operationId": "CvController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/CV" } }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "List my generated CVs",
        "tags": ["cv"]
      }
    },
    "/api/cv/templates": {
      "get": {
        "operationId": "CvController_getTemplates",
        "parameters": [],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "List available CV templates",
        "tags": ["cv"]
      }
    },
    "/api/cv/{id}": {
      "get": {
        "operationId": "CvController_findOne",
        "parameters": [
          { "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CV" } } }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Get CV detail",
        "tags": ["cv"]
      },
      "patch": {
        "operationId": "CvController_update",
        "parameters": [
          { "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/UpdateCvDto" } }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CV" } } }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Update CV content and name",
        "tags": ["cv"]
      },
      "delete": {
        "operationId": "CvController_remove",
        "parameters": [
          { "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CV" } } }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Delete a CV",
        "tags": ["cv"]
      }
    },
    "/api/cv/{id}/tailor": {
      "post": {
        "operationId": "CvController_tailor",
        "parameters": [
          { "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/TailorCvDto" } }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CV" } } }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Re-tailor an existing CV for a new job",
        "tags": ["cv"]
      }
    },
    "/api/cv/{id}/versions": {
      "get": {
        "operationId": "CvController_getVersions",
        "parameters": [
          { "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/CvVersion" } }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Get version history of a CV",
        "tags": ["cv"]
      }
    },
    "/api/cv/{id}/versions/{versionId}/restore": {
      "post": {
        "operationId": "CvController_restoreVersion",
        "parameters": [
          { "name": "id", "required": true, "in": "path", "schema": { "type": "string" } },
          { "name": "versionId", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CV" } } }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Restore a previous version",
        "tags": ["cv"]
      }
    },
    "/api/cv/{id}/html": {
      "get": {
        "operationId": "CvController_getHtml",
        "parameters": [
          { "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Get CV HTML for client-side rendering/printing",
        "tags": ["cv"]
      }
    },
    "/api/cover-letter/generate": {
      "post": {
        "operationId": "CoverLetterController_generate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/GenerateCoverLetterDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/CoverLetter" } }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Generate a Cover Letter",
        "tags": ["cover-letter"]
      }
    },
    "/api/cover-letter": {
      "get": {
        "operationId": "CoverLetterController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/CoverLetter" }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "List my cover letters",
        "tags": ["cover-letter"]
      }
    },
    "/api/cover-letter/{id}": {
      "get": {
        "operationId": "CoverLetterController_findOne",
        "parameters": [
          { "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/CoverLetter" } }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Get Cover Letter detail",
        "tags": ["cover-letter"]
      },
      "patch": {
        "operationId": "CoverLetterController_update",
        "parameters": [
          { "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateCoverLetterDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/CoverLetter" } }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Update Cover Letter",
        "tags": ["cover-letter"]
      },
      "delete": {
        "operationId": "CoverLetterController_remove",
        "parameters": [
          { "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Delete Cover Letter",
        "tags": ["cover-letter"]
      }
    },
    "/api/follow-up/generate": {
      "post": {
        "operationId": "FollowUpController_generate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/GenerateFollowUpDto" } }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/FollowUp" } }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Generate a follow-up email draft",
        "tags": ["follow-up"]
      }
    },
    "/api/follow-up/{id}": {
      "patch": {
        "operationId": "FollowUpController_update",
        "parameters": [
          { "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/UpdateFollowUpDto" } }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/FollowUp" } }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Update a follow-up draft content",
        "tags": ["follow-up"]
      }
    },
    "/api/follow-up/send": {
      "post": {
        "operationId": "FollowUpController_send",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/SendFollowUpDto" } }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/FollowUp" } }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Send or schedule a follow-up email",
        "tags": ["follow-up"]
      }
    },
    "/api/follow-up/track/{token}": {
      "get": {
        "operationId": "FollowUpController_track",
        "parameters": [
          { "name": "token", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "" } },
        "summary": "Track email open",
        "tags": ["follow-up"]
      }
    },
...
  },
  "info": {
    "title": "AI Job Application Context API",
    "description": "The AI Job Application API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": { "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" } },
    "schemas": {
...
      "Job": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "source": { "type": "string" },
          "title": { "type": "string" },
          "company": { "type": "string" },
          "location": { "type": "string" },
          "logoUrl": { "type": "string" },
          "companyAddress": { "type": "string" },
          "companySize": { "type": "string" },
          "workingTime": { "type": "string" },
          "companyType": { "type": "string" },
          "salaryMin": { "type": "number" },
          "salaryMax": { "type": "number" },
          "currency": { "enum": ["VND", "USD", "EUR", "JPY", "Other"], "type": "string" },
          "jobType": {
            "enum": [
              "Full-time",
              "Part-time",
              "Contract",
              "Freelance",
              "Internship",
              "Remote",
              "Hybrid",
              "Other"
            ],
            "type": "string"
          },
          "experienceLevel": {
            "enum": [
              "Intern",
              "Fresher",
              "Junior",
              "Middle",
              "Senior",
              "Lead",
              "Manager",
              "Director",
              "Not specified"
            ],
            "type": "string"
          },
          "level": { "type": "string" },
          "category": { "type": "string" },
          "categories": { "type": "array", "items": { "type": "string" } },
          "education": {
            "enum": [
              "Không yêu cầu",
              "THPT",
              "Trung cấp",
              "Cao đẳng",
              "Đại học",
              "Thạc sĩ",
              "Tiến sĩ",
              "Khác"
            ],
            "type": "string"
          },
          "city": {
            "enum": [
              "An Giang",
              "Bà Rịa - Vũng Tàu",
              "Bắc Giang",
              "Bắc Kạn",
              "Bạc Liêu",
              "Bắc Ninh",
              "Bến Tre",
              "Bình Định",
              "Bình Dương",
              "Bình Phước",
              "Bình Thuận",
              "Cà Mau",
              "Cần Thơ",
              "Cao Bằng",
              "Đà Nẵng",
              "Đắk Lắk",
              "Đắk Nông",
              "Điện Biên",
              "Đồng Nai",
              "Đồng Tháp",
              "Gia Lai",
              "Hà Giang",
              "Hà Nam",
              "Hà Nội",
              "Hà Tĩnh",
              "Hải Dương",
              "Hải Phòng",
              "Hậu Giang",
              "Hòa Bình",
              "Hồ Chí Minh",
              "Hưng Yên",
              "Khánh Hòa",
              "Kiên Giang",
              "Kon Tum",
              "Lai Châu",
              "Lâm Đồng",
              "Lạng Sơn",
              "Lào Cai",
              "Long An",
              "Nam Định",
              "Nghệ An",
              "Ninh Bình",
              "Ninh Thuận",
              "Phú Thọ",
              "Phú Yên",
              "Quảng Bình",
              "Quảng Nam",
              "Quảng Ngãi",
              "Quảng Ninh",
              "Quảng Trị",
              "Sóc Trăng",
              "Sơn La",
              "Tây Ninh",
              "Thái Bình",
              "Thái Nguyên",
              "Thanh Hóa",
              "Thừa Thiên Huế",
              "Tiền Giang",
              "Trà Vinh",
              "Tuyên Quang",
              "Vĩnh Long",
              "Vĩnh Phúc",
              "Yên Bái",
              "Toàn quốc",
              "Khác"
            ],
            "type": "string"
          },
          "isBranded": { "type": "boolean" },
          "tags": { "type": "array", "items": { "type": "string" } },
          "quantity": { "type": "number" },
          "gender": { "enum": ["Nam", "Nữ", "Không yêu cầu"], "type": "string" },
          "deadline": { "format": "date-time", "type": "string", "nullable": true },
          "allowance": { "type": "string" },
          "equipment": { "type": "string" },
          "industry": {
            "enum": [
              "IT - Phần mềm",
              "Tài chính - Ngân hàng",
              "Kinh doanh - Marketing",
              "Sản xuất",
              "Giáo dục",
              "Y tế",
              "Bán lẻ",
              "Vận tải - Logistics",
              "Xây dựng",
              "Khác"
            ],
            "type": "string"
          },
          "salary": { "type": "string" },
          "description": { "type": "string" },
          "requirements": { "type": "string" },
          "benefits": { "type": "string" },
          "skills": { "type": "array", "items": { "type": "string" } },
          "originalData": { "type": "object" },
          "expired": { "type": "boolean" },
          "externalId": { "type": "string", "nullable": true },
          "url": { "type": "string" },
          "postedAt": { "format": "date-time", "type": "string", "nullable": true },
          "createdAt": { "format": "date-time", "type": "string" },
          "isVerified": { "type": "boolean" },
          "isAlertSent": { "type": "boolean" },
          "status": { "enum": ["Pending", "Approved", "Rejected"], "type": "string" },
          "searchVector": { "type": "string" },
          "expiresAt": { "format": "date-time", "type": "string", "nullable": true },
          "contentHash": { "type": "string", "nullable": true },
          "updatedAt": { "format": "date-time", "type": "string" }
        },
        "required": [
          "id",
          "source",
          "title",
          "company",
          "location",
          "logoUrl",
          "companyAddress",
          "companySize",
          "workingTime",
          "companyType",
          "salaryMin",
          "salaryMax",
          "currency",
          "jobType",
          "experienceLevel",
          "level",
          "category",
          "categories",
          "education",
          "city",
          "isBranded",
          "tags",
          "quantity",
          "gender",
          "deadline",
          "allowance",
          "equipment",
          "industry",
          "salary",
          "description",
          "requirements",
          "benefits",
          "skills",
          "originalData",
          "expired",
          "externalId",
          "url",
          "postedAt",
          "createdAt",
          "isVerified",
          "isAlertSent",
          "status",
          "searchVector",
          "expiresAt",
          "contentHash",
          "updatedAt"
        ]
      },
      "CV": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "userId": { "type": "string" },
          "user": { "$ref": "#/components/schemas/User" },
          "jobId": { "type": "string" },
          "job": { "$ref": "#/components/schemas/Job" },
          "name": { "type": "string" },
          "content": { "type": "object" },
          "template": { "type": "string" },
          "score": { "type": "number" },
          "createdAt": { "format": "date-time", "type": "string" }
        },
        "required": [
          "id",
          "userId",
          "user",
          "jobId",
          "job",
          "name",
          "content",
          "template",
          "score",
          "createdAt"
        ]
      },
...
      "RefreshToken": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "tokenHash": { "type": "string" },
          "userId": { "type": "string" },
          "user": { "$ref": "#/components/schemas/User" },
          "isRevoked": { "type": "boolean" },
          "expiresAt": { "format": "date-time", "type": "string" },
          "replacedBy": { "type": "string" },
          "createdAt": { "format": "date-time", "type": "string" }
        },
        "required": [
          "id",
          "tokenHash",
          "userId",
          "user",
          "isRevoked",
          "expiresAt",
          "replacedBy",
          "createdAt"
        ]
      },
      "CoverLetter": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "userId": { "type": "string" },
          "user": { "$ref": "#/components/schemas/User" },
          "jobId": { "type": "string" },
          "job": { "$ref": "#/components/schemas/Job" },
          "content": { "type": "string" },
          "tone": { "type": "string" },
          "language": { "type": "string" },
          "createdAt": { "format": "date-time", "type": "string" }
        },
        "required": [
          "id",
          "userId",
          "user",
          "jobId",
          "job",
          "content",
          "tone",
          "language",
          "createdAt"
        ]
      },
...
      "UserCredits": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "userId": { "type": "string" },
          "balance": { "type": "number" },
          "lastRefillDate": { "format": "date-time", "type": "string", "nullable": true },
          "updatedAt": { "format": "date-time", "type": "string" },
          "user": { "$ref": "#/components/schemas/User" }
        },
        "required": ["id", "userId", "balance", "lastRefillDate", "updatedAt", "user"]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "email": { "type": "string" },
          "passwordHash": { "type": "string", "nullable": true },
          "isVerified": { "type": "boolean" },
          "isBanned": { "type": "boolean" },
          "verificationToken": { "type": "string", "nullable": true },
          "resetPasswordToken": { "type": "string", "nullable": true },
          "resetPasswordExpires": { "format": "date-time", "type": "string", "nullable": true },
          "role": { "enum": ["user", "admin"], "type": "string" },
          "googleId": { "type": "string", "nullable": true },
          "githubId": { "type": "string", "nullable": true },
          "appleId": { "type": "string", "nullable": true },
          "avatarUrl": { "type": "string", "nullable": true },
          "createdAt": { "format": "date-time", "type": "string" },
          "updatedAt": { "format": "date-time", "type": "string" },
          "profile": { "$ref": "#/components/schemas/Profile" },
          "trackers": { "type": "array", "items": { "$ref": "#/components/schemas/JobTracker" } },
          "cvs": { "type": "array", "items": { "$ref": "#/components/schemas/CV" } },
          "refreshTokens": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/RefreshToken" }
          },
          "coverLetters": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/CoverLetter" }
          },
          "jobAlert": { "$ref": "#/components/schemas/JobAlert" },
          "credits": { "$ref": "#/components/schemas/UserCredits" }
        },
        "required": [
          "id",
          "email",
          "passwordHash",
          "isVerified",
          "isBanned",
          "verificationToken",
          "resetPasswordToken",
          "resetPasswordExpires",
          "role",
          "googleId",
          "githubId",
          "appleId",
          "avatarUrl",
          "createdAt",
          "updatedAt",
          "profile",
          "trackers",
          "cvs",
          "refreshTokens",
          "coverLetters",
          "jobAlert",
          "credits"
        ]
      },
...
      "GenerateCvDto": {
        "type": "object",
        "properties": {
          "jobId": { "type": "string", "format": "uuid" },
          "template": { "type": "string" }
        },
        "required": ["jobId"]
      },
      "TailorCvDto": {
        "type": "object",
        "properties": {
          "jobId": {
            "type": "string",
            "format": "uuid",
            "description": "The target job ID to tailor this CV for"
          }
        },
        "required": ["jobId"]
      },
      "UpdateCvDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "description": "New name for the CV" },
          "content": { "type": "object", "description": "Updated CV content" },
          "template": { "type": "string", "description": "Template to use" }
        }
      },
      "CvVersion": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "cvId": { "type": "string" },
          "cv": { "$ref": "#/components/schemas/CV" },
          "versionNumber": { "type": "number" },
          "content": { "type": "object" },
          "createdAt": { "format": "date-time", "type": "string" }
        },
        "required": ["id", "cvId", "cv", "versionNumber", "content", "createdAt"]
      },
      "GenerateCoverLetterDto": {
        "type": "object",
        "properties": {
          "jobId": { "type": "string", "format": "uuid" },
          "tone": {
            "type": "string",
            "description": "professional | concise | friendly",
            "default": "professional"
          },
          "language": { "type": "string", "description": "en | vi", "default": "en" },
          "fullName": { "type": "string" },
          "skills": { "type": "array", "items": { "type": "string" } },
          "experience": { "type": "array", "items": { "type": "object" } }
        },
        "required": ["jobId"]
      },
      "UpdateCoverLetterDto": {
        "type": "object",
        "properties": {
          "content": { "type": "string" },
          "tone": { "type": "string" },
          "language": { "type": "string" }
        }
      },
      "GenerateFollowUpDto": {
        "type": "object",
        "properties": {
          "jobId": { "type": "string", "format": "uuid" },
          "type": {
            "enum": ["after_apply", "after_interview", "check_in", "offer"],
            "type": "string",
            "default": "after_apply"
          },
          "tone": {
            "type": "string",
            "description": "Tone of the email (professional, casual, enthusiastic)",
            "example": "professional"
          }
        },
        "required": ["jobId"]
      },
      "FollowUp": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "userId": { "type": "string" },
          "user": { "$ref": "#/components/schemas/User" },
          "jobId": { "type": "string" },
          "job": { "$ref": "#/components/schemas/Job" },
          "type": {
            "enum": ["after_apply", "after_interview", "check_in", "offer"],
            "type": "string"
          },
          "status": { "enum": ["draft", "scheduled", "sent"], "type": "string" },
          "content": { "type": "string" },
          "trackingToken": { "type": "string" },
          "openedAt": { "format": "date-time", "type": "string" },
          "scheduledAt": { "format": "date-time", "type": "string" },
          "createdAt": { "format": "date-time", "type": "string" }
        },
        "required": [
          "id",
          "userId",
          "user",
          "jobId",
          "job",
          "type",
          "status",
          "content",
          "trackingToken",
          "openedAt",
          "scheduledAt",
          "createdAt"
        ]
      },
      "UpdateFollowUpDto": {
        "type": "object",
        "properties": { "content": { "type": "string", "example": "Subject: ...\n\nBody: ..." } },
        "required": ["content"]
      },
      "SendFollowUpDto": {
        "type": "object",
        "properties": {
          "followUpId": { "type": "string", "format": "uuid" },
          "scheduledAt": { "type": "string", "description": "If provided, schedules the email." }
        },
        "required": ["followUpId"]
      },
...
    }
  }
}